requirements.txt
appdirs==1.4.4
asgiref==3.2.10
cfgv==3.2.0
distlib==0.3.1
Django==3.1.2
django-qrcode==0.3
django-sequences==2.6
filelock==3.0.12
identify==1.5.10
nodeenv==1.5.0
Pillow==8.0.0
pre-commit==2.9.3
PyMySQL==0.10.1
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.67
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
virtualenv==20.2.2
